MAJOR REFACTOR!
- Refactor renamed frontend variables to use proper naming conventions.
    - Renamed all underscores to dashes and snake_case to mixedCase as well as other respective JavaScript, HTML and CSS naming conventions.
- Replaced hardcoded HTML DOM Element identifier strings with constants wherever possible.
- Some code cleanup also happened intermittently throughout the process.
- Added more FIXMEs across codebase.

The use of constants instead of hardcoded strings will make renaming elements a lot easier (and less painful), as well as getting a more clear overview over what are actual strings and what are HTML DOM element identifiers/classes.

All identifiers, variables and constants now follow a more logical and intuitive naming scheme. This makes it better to work with and lets you know which is which or let's you guess what a likely identifier is named without having to look it up, repeatedly.